+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Fixed
+
+- Fix edge case when empty suffix `""` given to processor
+
 ## [0.3.0] - 2024-09-10
 
 ✨ This release is an exhaustive package refacto, making ColPali more modular and easier to use.